Full index recovery runs in a single thread that is kicked off by quartz.  This prevents multiple quartz threads from attempting to do the same work.
More control has been added over the behaviour of the index recovery.  By default, it can now run in a live server and will quit once the indexes are up to date, leavign the normal server processes to keep the indexes correct.  Additionally, it uses the L2 cache by default - also for the live situation.