Fix for WCM-934.
Now when link validation poller is disabled by setting poll interval to 0,
the "Check Links" control uses the last validated snapshot in staging
without throwing an NPE.

That said, I think it's a very poor idea to allow users to validate
links when the linkvalidation service is turned off because the delta 
between a workarea and the last validated snapshot of staging can become 
extremely large.  

Consider this:   If you have N users validating their in individual 
                 workareas, you can  end up spawning a huge & concurrent 
                 set of validation tasks that might need to validate 
                 hundreds of thousands of assets *each* !

The linkvalidation service amortizes link validation across all the workareas,
and makes the remaining cost in staging incremental.  The linkvaldation service
makes preliminary validation scalable.  Without it, it's not scalable.

Given that, I think the GUI ought disabling the "Check Links" control entirely 
when the linkvalidation service is deactivated.  All this fix does is make
it work (but not in a way that scales) when it's off.