Initial check in.
This project is a single, extensible framework for building dynamic web sites.  It includes a rich object model for handling scoped components, pages, and templates.  It manages the assembly of site construction objects and provides a solid runtime for presenting the site to the end user.

The code promoted here is a first-pass for the server-side components of what was formerly known as "Alfresco Dynamic Website".  That project has now been split roughly into UI and Server Side parts.  This project represents the latter.

This framework forms the foundation for Alfresco Web Client 3.0 as well as Alfresco Network - two  applications that make sure of this frameworks pluggability features.