Fixed ETWOONE-201: Better logging and configurability for RetryingTransactionHelper
log4j.logger.org.alfresco.repo.transaction.RetryingTransactionHelper=...
 - DEBUG dumps full stack traces for retry causes
 - INFO dumps one line summary of retry causes

Transaction retry behaviour can be set with the following properties with defaults:
 - server.transaction.max-retries=20
 - server.transaction.min-retry-wait-ms=100
 - server.transaction.max-retry-wait-ms=2000
 - server.transaction.wait-increment-ms=100

This puts lower and upper bounds on the retry pauses, as well as adjusts the amount of time
added to retry wait with each iteration.  The defaults are shorter than they were (500ms+count*500ms)
but the random nature could give retries of a few milliseconds.