Upgraded to commons-pool-1.4
The concurrency model is much better in this pool implementation.  Profiling revealed
that the checking in and out of DB connections was slowing things down as it synchronized
around a piece of code that included the cleanup of the DB connection.  Now, the DB
connection cleanup still takes a while but at least all threads do that in their own time
without blocking.