ETHREEOH-291
The My Sites dashlet displayed the first site incorrectly in a 2 column layout.
The reason for this was because the my sites dashlet (and several others) use the "detail-list-item" selector combined with "first" & "last" selector.
Since the dashlets are placed in columns using YUI css sleectors like "yui-gc" and "yui-gd" which also use "first" & "last" selectors, the width property was in some situation overriden in FF3, IE6 & IE7.
The "first" & "last" selectors for the "detail-list-item" selector have now been refactored to be called "first-item" & "last-item" in all dashlets.