Added a sequenceNumber attribute to the model abstract class. I also added getter and setter methods for this attribute. This attribute will be assigned by the parent to it's child nodes. The nodes will then use this value in their title div. This is for child nodes that do not have names, like rows and columns.
Modified the way the id attribute is generated for rows, columns, and regions. Now, I use a utility function in /includes/utils.js to generate an id.

Added a height and width attribute to the template json. Also modified the create and edit wizards for this value to appear when dealing with templates.