Fix ETWOTWO-973 - Enable Logging of Leaked Transactions By Default.
A leaked transaction (i.e. a begin is not matched by a commit or rollback) is a frequent reason for requiring a reboot of the server due to lost connections or other resources.

It was possible before to manually switch on logging of leaked transactions (by setting log4j.logger.org.alfresco.util.transaction.SpringAwareUserTransaction.trace=DEBUG).  Note: this has an associated performance penalty, so the problem is that this manual step is re-active and generally too late, as it requires the server to be rebooted and the use case to be reproduced.

This fix means the server will now log leaked transactions by default without manual configuration meaning any code out there that's badly managing transactions will now be exposed in logs. To get around the performance issue, the tracing of the leaked transaction call stack (very useful for locating bad code) is now switched on automatically after the first detection of a leaked transaction. Performance suffers from that point on, but your server is hosed anyway. 

Leaking transactions should no longer be the hidden cause of server issues, wherever the client code comes from.