Fix ETWOTWO-1058: Hibernate exception while concurrently submitting from and updating same user sandbox.
Root cause of issue was that jBPM job handling (which is responsible for async submit) was not working as should in our Alfresco integrated version of jBPM.  This was due to the fact that sessions/transactions may not (under some circumstances) be shared between Alfresco and jBPM.  This meant that jBPM job state and Alfresco state became out of sync, that is, the job thought it should complete and destroy itself while the Alfresco submit failed due concurrent update to user sandbox.  Alfresco's retrying transaction handler kicked in to retry the job, but the job was no longer there.

Ideally, the retry transaction handler should handle the case of failed concurrent update to the user sandbox, even inside an async jBPM job.

The fix involved:
- re-configuring jbpm so it does not attempt to create any sessions or start transactions
- all transaction boundaries are controlled by Alfresco and inherited by jBPM
- re-worked jBPM job queue to take above into account
- job execution now wrapped carefully in Retrying transaction handler
- improved job scheduler shutdown handling
- worked around jBPM bug (now exposed due to above) where job is deleted in multiple ways during transaction causing hibernate to barf

So, now if a job fails, it's first protected by Alfresco's retrying transaction handler, and failing that, is retried (at a later time) according to jBPM own job handling rules.

This fix should improve all workflow scenarios, not just wcm submit as transactions are now always shared between Alfresco and jBPM, as well as take less db resources during workflow processing.

Successfully executed the jmeter test.  Each submit carries only the one item it should, and 30 submits takes place in all.  Concurrent exceptions will appear in logs, but are expected if concurrent modification on a user sandbox take place.  The retrying transaction handler compensates for that, so is not a bug.