Support for variable assignment and replacement in upgrade scripts
The --ASSIGN: directive allows a value to be extracted from the resultset of the SQL
statement immediately following.  So, the following script:
   --ASSIGN:x=blah
   select 12345 as blah;
   create table avm_aspects_${x}
   (
        id bigint not null auto_increment,
        primary key (id)
   ) type=InnoDB;
will actually execute the following SQL:
   select 12345 as blah;
   create table avm_aspects_12345 (id bigint not null auto_increment, primary key (id)) type=InnoDB;
Notice that direct string replacement of the ${x} is done before the SQL statement is prepared,
allowing modification of column definitions (e.g. setting auto-increment counters) based on values
read earlier in the script.

Once a variable has been assigned, it is available for the whole script.  All variable assignments
are only available in the script in which they are declared.