ETHREEOH-2566 : Enterprise 3.x / Versioning is lost on Content after the Import and Export process of a Space
Export of the version history of the versioned nodes was implemented.
Import of the version history was implemented.
Method getTargetAssocs of Node2ServiceImpl class was modified to provide .rootVersion. associations for the versions.
VersionHistory interface was extended with method:    public NodeRef getVersionHistoryNodeRef(NodeRef nodeRef); to provide 
access to the node.s version history for export.
ExporterComponentTest and ImporterComponentTest were modified . checks of the export/import of versioned nodes were added.