ETHREEOH-1951: when versionable aspect is active, using the Microsoft Word option "always create a backup copy" leads to document versions loss
We add exstension *.wbk to the tempFileMarkerInterceptor, and this prevent of creation extra version.
Then we change ContentDiskDriver.renameFile() method.
When destination file has status DeleteOnClose, then the standard alfresco behaviour was:
1. Recover this file.
2. Set content from src file.
3. Delete src file.
4. Remove status of src file.
We change fourth point. When src file is deleted we set "DeleteOnClose" status to it(with expiration time in 1 minute).
And when MS word try to create this file, it will recover from archive store and version history will stay.