ETHREEOH-3193: 'capitalize' in output path pattern works differently for templates (vs. XSDs)
This fix makes output paths perform same behaviour using the same freemarker string between the file output path and the rendition 
output path.
Note! The previous behaviour may also look correct, because the initial name of the output paths of the renditions is the output 
name of the form already PROCESSED by FreeMarker.