ETHREEOH-1211: Can't See Images in TinyMCE
This issue  is fixed for IE. 

But latest changes in xforms.js raise new bug in FF. 
When user try to edit webcontent the image dissapeares. This issues are fixed too by this checkin.

Note! There is a conflict with ETHREEOH-2789: During fixing of ETHREEOH-2789, forced_root_block param in tineMCE become false. 

However it is strongly recomended don’t do this, because in this case there is no guarantee that editing will work correct. 
http://wiki.moxiecode.com/index.php/TinyMCE:Configuration/force_br_newlines
http://wiki.moxiecode.com/index.php/TinyMCE:FAQ#TinyMCE_produce_BR_elements_on_enter.2Freturn_instead_of_P_elements.3F

We would like recommend to revisite ETHREEOH-2789.