ETHREEOH-3239: Version number doesn't change after modifying document using 'Edit Online' feature.
- Upload web script was changed for support version creation. The cm:autoVersion and cm:initialVersion properties were changed 
to true.
Note! It consists of two parts. The first is a version creation itself. This part was successfully implemented, but the 
another part is the UI component refresh after the editing of the document ends. This hasn’t been implemented because the only 
way to update browser window in such case is the timer event, but this functionality may became an overkill in case of many 
opened windows count.