ETHREEOH-3454: Port enterprise upgrade scripts for ContentStoreCleaner changes
The parameter map for alf_content_url table mapping was added. The index creation on the alf_content_url table was changed for Oracle dialect. The 
alf_content_url table creation was changed for DB2Dialect. Also index creation was changed for  alf_content_url table in DB2Dialect and in 
AlfrescoSQLServerDialect.