ETHREEOH-3187: Creating Web Content based on Web Form with most elements-read-only/default failed
The value of the returned from XSD doesn’t contain milliseconds part on JBoss. And as a result javascript time parsing fails. The new 
additional time pattern without milliseconds was added to parsing. So, if the time cannot be parsed against "HH:mm:ss.S" then it is 
parsed against "HH:mm:ss".