ETHREEOH-3473: It is impossible to upload and create web content
The reason of bulk upload fails was in “update_AVMNodeProperty” sql statement. However we haven’t find the way to make it working for DB2 and 
PostgreSql in one form. The decision was to move this statement to dialect specific ‘avm-insert-SqlMap.xml’ and change it for PostgreSql. Running 
unit-tests for repository project we have founded out a mistake in ‘audit-insert-SqlMap.xml’ (postgresql) and also fixed it.