ETHREEOH-3850: Content Manager unable to edit content item if there is a lock on the generated rendition
The action logic of the private boolean hasAccess(String webProject, NodeRef webProjectRef, String avmPath, String user) method has 
been changed in org.alfresco.repo.avm.locking.AVMLockingServiceImpl class.
The previous algorithm that granted access for user to avmPath was the following:
1.	If user is admin.
2.	If lock doesn’t exists for avmPath.
3.	Store name for avmPath contains "--workflow".
4.	lock.getStore equals Store name for avmPath and user is CONTENT_MANAGER.
5.	lock.getStore equals Store name for avmPath and user is lock owner.

In case of situation described in the issue, the store for lock (webProject--user) and the store for 
avmPath(webProject--user--preview) are not the same. That’s why the access to the content for the user isn’t granted even to user with 
role CONTENT_MANAGER and the lock owned. We would like to suggest the following algorithm to grant access:
1.	If user is admin.
2.	If lock doesn’t exists for avmPath.
3.	Is user a lock owner.
4.	Store name for avmPath contains "--workflow".
5.	lock.getStore equals Store name for avmPath and user is CONTENT_MANAGER.