ETHREEOH-3979: Web Forms: Inline images within TinyMCE controls aren't displayed correctly
When user upload image it goes to the previw store(not main user store), so we edit AVM_WEBAPP_URL for pointing to the preview.
alfresco_TinyMCE_urlconverter_callback was also changed. There is no need to remove AVM_WEBAPP_URL here. It will be removed during 
saving content in _tinyMCE_blurHandler.