ETHREEOH-3861: Error occurs when you try to change name of locked file
The cause of the issue is incorrect avm path returned for a node when MT is enabled. The “UNKNOWN/UNKNOWN” path is detected and the method 
returns null in this case. It fixes the issue with UI error displayed.
This fix also includes default value for TenantRoutingFileContentStore#applicationEventPublisher property in 
mt-contentstore-context.xml.sample. It is necessary for successful enabling MT with renaming mentioned file to mt-contentstore-context.xml.