ETHREEOH-3850: Content Manager unable to edit content item if there is a lock on the generated rendition
The previous check-in is not completely suitable CHK-10864. We got it during running unit tests we made. Thus, we have made a new fix that 
doesn't return access denied if the lock store and the target store is user store and it's preview.
This check-in contains new fix and unit tests.