New 'Site Creators' group, membership of this group will define that you have the authority to create sites.
Update the create site service Web Script (repo) tier to check for
membership of the Site Creators group

Update the web tier (Sites dashlet and toolbar) to check if the user is a
member of the Site Creator group and show the create site link as
appropriate.

Upgrades


- Patch to create the Site Creators group


- Config in alfresco-global.properties to enable authority based site
creation, this should be disabled by default so that existing installs will
have the existing behavior by default.


Customers will need to explicitly enable the authority based creation by
changing the default config.

Unit tests for the above.