ETHREEOH-4044: External Authentication Subsystem does not perform user mapping for WebDAV requests
The changes listed in the bug comments have been made.
Additional notes on RemoteUserMapper interface:
It is declared in the Web-Client project and is invisible to the Remote API project but it is necessary for WebDAV authentication 
filters.
So, we see two approaches to obtain “user mapper” in the WebDAV:
1.	Make duplication of the external functionality within Remote API. I.e. create the same or similar RemoteUserMapper interface, 
DefaultRemoteUserMapper class and make duplication of the properties for external authentication. We think this approach is not 
suitable because it is not flexible.
2.	Move RemoteUserMapper interface into Remote API project. The sources of the Remote API are visible for the Web-Client 
project, so, it can be used for both web-client authentication and WebDAV authentication. Using this approach DefaultRemoteUserMapper 
implementation can be accessible via Spring in the RemoteAPI.
The second approach has been implemented.