ETHREEOH-4157: Alfresco failing to startup with Websphere 7 and IBM JDK 1.6_SR6
The cause of the issue is the changes in the java.util.LinkedHashMap in the IBM JDK 1.6_SR6 (WAS SDK 7 FP7) update. This class is 
used by Ehcache in the net.sf.ehcache.store.LruMemoryStore$SpoolingLinkedHashMap.

We have added a NPE check in the last one like it was done by guys from Ehcache (https://jira.terracotta.org/jira/browse/EHC-533)