ETHREEOH-4031: Unable to delete space that has a rule attached to it
The more deeply investigation has shown that this is an issue with HibernateNodeDaoServiceImpl.parentAssocsCache. This issue is 
raised only if inbound rule is created for a target folder, and it is raised when DbNodeServiceImpl.pullNodeChildrenToSameStore 
method is trying to move the system ruleFolder to the same store as target space. E.g. we have “Company Home > TestSpace”  space 
and an inbound rule attached to the TestSpace. In this case all inbound and outbound rules are correctly being invoked in the 
appropriate policies. Then, when the TestSpace has been already moved into the archive store, the pullNodeChildrenToSameStore 
method makes move of the system ruleFolder to the same store as TestSpace. The newChildNodePair = 
nodeDaoService.moveNodeToStore(“ruleFolder”, “archiveStore”) is invoked. This call updates the 
HibernateNodeDaoServiceImpl.parentAssocsCache for the ruleFolder.id, BUT IT DOESN’T for rule nodes contained in this folder. 
Further invocation of the invokeOnCreateNode in the DbNodeServiceImpl.pullNodeChildrenToSameStore triggers the rules to be 
executed. In this case RuleServiceImpl. getOwningNodeRefRuleImpl returns OLD NON-EXISTENT primary parent from 
this.nodeService.getPrimaryParent(ruleNodeRef).getParentRef() call. I.e. the OLD value from cache. The next invocation of the 
this.nodeService.getPrimaryParent(systemFolder).getParentRef() with non-existent systemFolder fails.  

We would like to suggest to remove old child nodes from HibernateNodeDaoServiceImpl.parentAssocsCache within 
nodeDaoService.moveNodeToStore(“ruleFolder”, “archiveStore”)  method.