ETWOTWO-738: Deleting space with "execute a script" rule fails
The cause of the issue is out-of-date storeRef of the rule that is going to be executed before transaction commit. The rule object 
that has been bound to the AlfrescoTransactionSupport in the RuleServiceImpl.addRulePendingExecution become out-of-date after moving 
itself to the archive store. This rule object is not being updated during further moving of the child rules and child actions, 
because this rule object is taken only once for the every content of the target folder  (the folder with rules attached). So, when 
the already bounded rule is taken from AlfrescoTransactionSupport in the RuleServiceImpl.executePendingRulesImpl during 
doBeforeCommit execution the action of the rule holds old storeRef of the rule that has been already moved. 

We would like to suggest to refresh rule taken from AlfrescoTransactionSupport. This fix includes this approach. As an alternative 
approach we can take a rule NodeRef by UUID from DAO layer but it takes extra effort, i.e. new methods and DI.

The 3.2 version can be fixed with this fix too.