Committing interim changes for Iteration 9, week one of Surevine work on Enhanced Security Checks to the slingshot project.
Addition of stub interfaces to comments, discussions, wikis and an early edition real interface to the html upload in the doc lib, allowing a user to enter ESC data when creating items.

This is an interim release in prep for QA and future merges.  Testing and internal QA has been minimal at best.