ETHREEOH-4182: ASR deployer fails to set the contentUrl of documents on the target system
The cause of the issue is incorrect handling of the output avm stream. The DeploymentServiceImpl.copyStream method is used in case of the 
content transfer. But in the DeploymentServiceImpl.deploySinglePush (line 534) somebody has enclosed this method into try-finally block and the 
use case when the content file is placed directly in the root folder works correctly. E.g. “main:/content.txt” in context of ASRDeploymentTest. 
But the case with nested files (placed in the directories “main:/a/content.txt”) dosn’t work. The files become empty because the copyStream is 
used further in the deploySinglePush method see lines 602,622,669.
So, we would like to suggest to flush and close output stream in the copyStream method. In this case the ASRDeploymentTest is working with 
commented out lines with content checking.
In the V3.1 the copyStream method out.close() invocation exists but in V3.2 doesn’t.