End of Iteration 9 Release.
Presentation changes throughout to allow ESLs to be applied on Wiki pages, Discussion groups and in the document library (only document library complete, the rest is stubs) and to render ESLs back out.

Code needs a thorough review - quality has been sacrificed for time to some extent and this needs improving before a merge back.

ESC unit tests and a good amount of manual testing done.  The purpose of this checkin is to establish a baseline for an interim client build - further testing will be performed at client site.