Merge Head to DEV/DM_Rendition_API r18542: Updated greenTheme inline with default theme CSS. Fix to I18N label name incorrectly specified in Edit Site Details dialog. Removal of hardcoded colour in Site Finder CSS.
r18545: Make it possible to request that the ApplicationContext disables "autoStart=true 
" services for when running unit tests 

r18548: Added "has-tag" action evaluator. 

r18551: Removed the need to supply a formId attribute on the 'markContent' tag and added 
 debug to the customer 'property' tag 

r18554: Added 'title' support. Allows a template developer to define a 'title' for the m 
arker, this will then be used in the quick edit button in the toolbar, the title 
 of the dialog/popup panel and as the alt and title attribute of the edit (penci 
l) image. 

A 'nested' property is also now passed in to the registerEditableContent functio 
n, this will be used to control the highlighting of content and/or markers. 

r18555: Updated hcBlack theme inline with default theme CSS. 

r18558: Highlighting of element when selected in quick links menu now respects the 'nest 
ed' setting of the marker