SAIL-70 Renditions are performed as system not as admin
Renditions are performed by the system user, not by the admin user.
This was a mistake in the previous check-in.

Related changes: StandardRenditionLocationResolverImpl now uses the fully authenticated user rather than the current user, when building the model (This person object may be removed from the model in future).
The TemplateModelHelper.buildDefaultModel now uses the fully authenticated user rather than RepositoryHelper.getPerson, which is the current user.
There are needed as renditions are no longer run by the currently logged in user.