ALF-1358: Deleting space with "execute a script" rule fails
If the rules’ store is not the same as action upon node ref’s store then we are checking the rule on existing in the original rule 
store. If the rule doesn’t exists in the original store, we are trying to test does the rule exists in the action upon nodeRef’s store. 
At now, if the rule doesn’t exist in both stores we do nothing and the exception is being thrown in the 
ScriptActionExecuter.executeImpl  - this.serviceRegistry.getRuleService().getOwningNodeRef(action).
The one of possible cases is to throw an Exception in the RuleServiceImpl.executePendingRule in case of the rule doesn’t exist in both 
stores. We need your suggestions what is the best way to handle such case. 

JUnit test for ETWOTWO-738 reproducing has been created.