SAIL-73 Renditions are updated on content update by default, on property change if configured.
Added a new optional property to all RenditionDefintions (on AbstractRenderingEngine) namely update-renditions-on-any-property-change. If this property is not specified or is false, then renditions will be automatically updated when any content property on the source node changes. If this parameter is true, then a change to any property on the source node will trigger rendition updates.
This is necessary for renditions based on a template/model where source node properties could appear in the rendition.
For renditions to be automatically updated in this way, they must have been saved.
A source node could have multiple renditions some of which update on content property change and some of which update on any property change. This behaviour is determined by the update-renditions-on-any-property-change parameter saved with the renditionDefintion

Reimplemented the internals of RenditionedAspect behaviour class completely.

Changed an existing template test case to test this update behaviour. I'm using the cm:title field on a source node and rendering it into the renditions using a template-based rendition. Then when I set other properties on that source node, the renditions may be updated as described.

Added some test data to the renditionTestTemplate.ftl