ALF-883: Rule is not triggered for a item which is restored from the bin
At the first look the rule should be triggered by CreateNodeRuleTrigger. But the javadoc says “The create node trigger will only be fired 
if the object is NOT a sub-type of content”, so the inbound rules for restored content are not triggered by this policy. Thus, we would 
like to suggest to trigger such rules in the special RestoreNodeRuleTrigger in compliance to CreateNodeRuleTrigger and be sure that its 
being invoked only when content is restored from recycle. It’s is invoked from DbNodeServiceImpl.restoreNode() method.