ALF-511: It is impossible to edit rendering templates for a web form
We are not able to reproduce this issue. But the visual analysis of the code shows that this issue is reproducible when there are several web 
forms with same name in the repository (see lucene query). So, the suggested fix, i.e. move property addition into for loop, looks reasonable 
in this case. We have implemented this fix in this check-in.