ALF-934: View Details does not refresh the doc properties when displaying multiple documents via Direct URL
When “showDocDetails” outcome is processed in the ExternalAccessServlet, the AlfrescoNavigationHandler is already invoking renderResponse 
phase of the JSF lifecycle at his end (private void goToView). But, the further invocation 
getServletContext().getRequestDispatcher(FACES_SERVLET + viewId).forward(req, res) via FacesContext starts the new JSF lifecycle which is 
started from restore view phase and the last view is displayed. This is the cause of the issue.
We suggest just to forward to the view itself without faces context for showDocDetails outcome, because the JSF lifecycle has been already 
finished by render response in the AlfrescoNavigationHandler.