ALF-2034: Error when deleting from the details view page
After deletion of a Node the AlfrescoNavigationHandler has returned user to the "Document details" view wich has raised the 
NullPointerException during displaying of properties of the deleted Node. The AlfrescoNavigationHandler uses a View Stack to navigate to 
previous views. The fix contains logic wich removes the "Document details" dialog from the View Stack in the 
DeleteContentDialog.finishImpl()