ALF-1523: Failed Kerberos SSO auth doesn't fail through, simply returns a blank page
When user doesn.t have his Kerberos ticket he will be redirected to the Alfresco login page and login will be allowed.
The same functionality was implemented for NTLM when authentication fails.
Also for Share application when NTLM authentication is used and it fails the user will be redirected to the Share login page