ALF-2011: Audit doesn't take into account CIFS authentication
Since each child of the CifsAuthenticator is not a Spring configured bean it has no Transaction interceptor. The Transaction wrapping 
functionality was added into the CifsAuthenticator.setCurrentUser() to fix a Transaction Synchronization issue after successful 
authentication of the user.
Several little corrections added into the InMemoryTicketComponentImpl to allow "Null user". "Missing ticket for null" exceptions will 
be thrown instead of the NullPointerException