ALF-1523: Failed Kerberos SSO auth doesn't fail through, simply returns a blank page
Functionality for Share version 3.3:
When user cancel NTLM authentication he is redirected to the login page and is able to login via WebClient.
Also we found NTLM authentication problem in Share 3.3 as described in:
http://forums.alfresco.com/en/viewtopic.php?f=8&t=26289
https://issues.alfresco.com/jira/browse/ALF-2420
Our first look shows that the problem probably occurs when RepositoryContainer class tries to authenticate.
Call to the “auth.authenticate()” on line 289 results in call to the BasicHttpAuthenticator instead of WebClientAuthenticator as it was 
in Alfresco 3.2.