ALF-2450: latin/utf-8 HTML file cannot be text-extracted.
The cause of the issue is following. During converting "text/html" to "text/plain" Alfresco founds 2 transformers: StringExtractingContentTransformer and HtmlParserContentTransformer. 
No one of them does support explicit transformations, and the fastest StringExtractingContentTransformer is used. So, HtmlParserContentTransformer is not used. 
So, the explicit transformation "text/html --> text/plain" was added to HtmlParserContentTransformer.