ALF-2507: Not able to email space users even if the user owns the space
The EmailSpaceUsersDialog.getUsersGroups() method is trying to retrieve all users that have access to the folder.
Also it is trying to get all users that have access to any of parent folders if InheritParentPermissions is true.
But if a user that perform email action doesn't have a permission that allows him to retrieve permissions of these (parent) folders, then exception is occurred and
the result is empty.
We wrapped a permission retrieving into RunAsWork from system user.