ALF-897: It is impossible to create content when default value selected in ContentHeadlineBackground field for intranet_rssi_landing_template web-form
This check-in contains the fix for ALF-897, ALF-791, ALF-2798.
ALF-897 – The extra field was added to the combo box. If a field is required, then user will get validation error.
If field is optional then it will be empty in result xml. The title  of extra field is "Please select value or leave blank for null." and can be configured
in webclient.properties within a key = "please_select".
ALF-791, ALF-2798 - Modified Schema2XForms to allow empty value if field is optional and has REGEXP.
Note! This fix doesn't work if field is optional attribute (i.e. <xs:attribute use="optional" ../>), see ALF-523 for details.