ALF-2430: System error happened during create web project if user is invited to Web Project space with Content Coordinator role
The ‘org_alfresco_module_dod5015’ module causes this issue to be replicable.
The problem related to the ACLEntryVoter bean with the ‘abstainFor’ property configured. Sometimes the vote may be performed for a NodeRef with Read Permissions that have not yet been set for current user (like in the situation described in this issue) and it is necessary to collect all info required for a vote with constantly acceptable rights.