ALF-872: TinyMCE Insert/edit Link dialog does not display properly in IE6
Drop-down and list boxes do not have a z-index property; these are window level controls. 
When you want to show a div in a page that contains these controls, you will see an overlapping problem.
This is a well-known problem with the IE 6 browser.
To solve this we just hiding form's divs until FilePickerWidget is undestroyed.