ALF-3579: Open the Details Page URL removes the header, navigation bar etc
The initial case for this issue was the fix for ALF-1010.
The changes were reverted for ALF-1010 and a new fix is applied. For now, ALF-1010 is fixed by this fix too.
The identifiers for modify action tag and details’ actions tag should be different in the details pages. For documents:
in the filelink-details.jsp and document-details.jsp. For spaces: in the space-details.jsp and spacelink-details.jsp.
Appropriate changes were made.