Workflow features 23 & 26, added "task-type" config evaluator.
This allows task ids i.e. "jbpm$34" to be passed to the config service and it look up the relevant form configuration for that type of task (similar to the node-type evaluator).

I've also added a first cut of the task form config for wf:adhocTask and wf:reviewTask. The form console can be used to test these, just set the item kind to "task" and the item id to the id of a workflow task.