ALF-3112: Property parameterTemplates not resolved correctly for SimpleTemplateActionDefinition
FixedPropertyPlaceholderConfigurer which is descendant of PropertyPlaceholderConfigurer. Also this class should have a resolvePlaceholder behavior from JndiPropertyPlaceholderConfigurer, but it is dependent of the boolean bean property "isJndi". If the bean is configured with this property is set to true, then jndi behaviour is used, else default behaviour is invoked. Then we are configuring abstract bean id="common-placeholder-configurer" of class FixedPropertyPlaceholderConfigurer and make "repository-properties", "shared-properties", "cacheStrategiesPlaceholderConfigurer" and other neccessary beans as descendants of "common-placeholder-configurer". E.g. <bean id="repository-properties" parent="common-placeholder-configurer"/>.