Second checkpoint for publishing. Nearly complete. Just publishing of deletes left to go.
Note that this commit includes a change to the content model. There is now a website type that takes on some of the responsibilities of the webroot type (hostname, port, and web context). A website now has a website node with a webroot node immediately beneath it named "root". This enables the root section to be published without overwriting the host/port config of the live edition of the site.