ALF-661: There is no way to determine the protocol, hostname and port from a javascript kicked off by JBPM
New class(ScriptUrls) that exposes the parameters of SysAdminParams was created.

Usual script:
New object("urls") that represent ScriptUrls was added to the data-model in ScriptServiceImpl.buildDefaultModel().
New script("append urls.js") that appends alfresco and share urls to file was added to "Data Dictionary\Scripts".

Scripts that are managed thro JBPM:
New object("urls") that represent ScriptUrls was added to the data-model in AlfrescoJavaScript.executeScript().
I add this object to data-model in AlfrescoJavaScript.executeScript() because adding it in JBPMEngine.startWorkflow() result in NotSerializableException(SubsystemProxyFactory is not completely serialized).