Fix for ALF-6192 alfresco share could not delete file with preview. This issue was ultimately caused by the incorrect use of a BeforeDeleteNode policy. As a workaround to ALF-4119 a beforeDeleteNode policy was defined in RenditionAspect.java. This policy ensured that when path/template-based renditions were deleted, that the parent-child association to the source node, which is a non-primary association, was removed. In doing so, future calls to renditionService.getRenditions() would not return renditions from the archive store.
However, this led to the permissions problem observed in this issue.

It is possible that running the BeforeDeleteNode policy as system could have resolved the problem.
However an alternative workaround for the deleted renditions problem has been implemented.
The policy has been removed and instead, the RenditionService filters the results.